Added a pugin that attempts to guess if a browser is a desktop.
Currently the desktop detection is done by DeviceAtals which requires a
licence it shows us as "n/a" if DA is not istalled.  This plugin
attempts to guess whether a browser is a desktop.  It's a simple bit of
user agent sniffing but should provide a stop-gap for this data point.